requirements.txt
Keras==2.2.4
res_manager==3.1.1
numpy==1.15.2
tensorflow_gpu==1.8.0
boto3==1.9.115
opencv_python==3.4.3.18
matplotlib==3.0.2
Algorithmia==1.1.4
Pillow==6.0.0
